Redesign suggestion admin UI into a tabbed layout
Consolidate the suggestions and (previously separate) categories admin
pages into one tabbed surface — Suggestions | Categories | Import & Export —
driven by a ?tab= query param so the household config cards still deep-link.

- Move the crowded export flow (an unbounded, unscrollable AI-prompt inside a
  dialog) into a dedicated, roomy Import & Export tab. The prompt is collapsed
  by default and rendered in a height-capped scroll box, so it can always be
  dismissed.
- Unify export/import into a single bundle file (recommendations + categories),
  orchestrated client-side over the existing endpoints; the separate category
  export/import is removed as redundant. Legacy single-half files still import.
- Make the duplicate-review action obvious: a full-text "Review duplicates"
  button with an explanatory banner instead of a bare toolbar icon.
- Remove the old categories route (no redirect shim) and point its config card
  at ?tab=categories.

Delete ExportSuggestionsDialog and the now-unused single-purpose export/import
hooks. Update unit tests and visual snapshots (baselines regenerated by the
workflow).
