+++ CLAUDE.md
@@ -171,6 +171,7 @@ This project uses SonarCloud for static analysis. Both backend (`vormadal_Anythi
 - **Suggestions (`ShoppingListRecommendation`) are list-scoped via a nullable `ShoppingListId`.** `null` = *shared* (surfaces in every shopping list's autocomplete — the legacy/global scope); a non-null value scopes the suggestion to that one list. This preserved backwards compatibility with **zero data migration**: pre-existing rows stayed `null` and thus visible everywhere. Query rule for a list's typeahead/management view: `HouseholdId == h && (ShoppingListId == listId || ShoppingListId == null)`. The item-add/recipe/food-plan seeding sites (`AddShoppingListItem`, `ShoppingListHelpers`, `AddRecipeToShoppingList`, `AddFoodPlanToShoppingList`) set `ShoppingListId` to the destination list, and their "already exists" check is scoped to `(list-or-shared)`. Uniqueness on `(HouseholdId, ShoppingListId, Name)` is enforced by **two partial unique indexes** — one `WHERE "ShoppingListId" IS NOT NULL`, one on `(HouseholdId, Name) WHERE "ShoppingListId" IS NULL` — instead of a single index with `.AreNullsDistinct(false)` (Postgres `NULLS NOT DISTINCT`, which requires PG15+ and broke the e2e/production deploy: the API crashes on `MigrateAsync()` before `app.RunAsync()` ever runs, which passes CapRover's build-only `verify-deploy.js` check but makes every request — including login — fail, surfacing as a `waitForURL` timeout in the auth setup e2e test). The two-partial-index form gives identical uniqueness semantics on any Postgres version. `GetShoppingListItems` resolves the name→category sort order **in memory** (preferring the list-specific row over the shared one) because a SQL join on name would now duplicate an item row when both a shared and a list-specific recommendation share the name. Admin: `DELETE /api/shopping-list-recommendations/by-list/{id}` clears only a list's own suggestions (shared ones stay); `GET /all` accepts `shoppingListId`/`sharedOnly`/`uncategorized`/`includeInSuggestions` filters. Recipe ingredient autocomplete stays household-wide (a recipe isn't a shopping list), so it passes no `shoppingListId`.
 - **CapRover's "deploy success" only means the image built — it never checks the app actually started, and its log API is undocumented.** `deployedVersion === latestVersion` on `/user/apps/appDefinitions` just means Docker finished building/handing off the image; the entrypoint can crash-loop forever (e.g. `MigrateAsync()` throwing on a `PendingModelChangesWarning`) and CapRover still reports the deploy as successful — this is the exact bug `.github/actions/verify-caprover-deploy` was built to catch, by additionally polling the app's own `/health` endpoint after the build finishes. Two more CapRover REST calls, found only by reading `caprover-api`'s (not `caprover-cli`'s — that package's `ApiManager` is an older, stripped copy) TypeScript source on npm, since none of this is in the public docs: build logs are `GET /user/apps/appData/<appName>` → `{ logs: { lines: [...] } }`; runtime container logs are `GET /user/apps/appData/<appName>/logs?encoding=hex` → `{ logs: "<hex>" }`, which is the **raw hex-encoded Docker multiplexed stdout/stderr stream** (8-byte frame header per chunk: 1 byte stream type, 3 reserved, 4-byte big-endian payload length) and must be decoded before it's readable — see `decodeDockerHexLogs` in the action's script. Both calls go through the existing `caprover api --path ... --method GET --data '...'` CLI invocation, no new auth mechanism needed.
 - **`caprover api` GET calls fail immediately (and misleadingly) if `--data` is omitted — always pass it.** The CLI's `data` option (`caprover/src/commands/api.ts`) has no `when: false` guard, so when `callCaproverApi()` in `verify-deploy.js` is called without a `data` argument (as `fetchAppDefinition()`/`printBuildLogs()` originally were), the CLI falls back to an interactive `? API data JSON string:` prompt. In non-interactive CI that prompt gets EOF and the process exits non-zero within ~1s, before any actual polling happens — surfacing as `caprover api GET /user/apps/appDefinitions failed: Call generic CapRover API [Experimental Feature]...`. `main()`'s catch then dumps `printDiagnostics()` (build + runtime logs) and exits 1, so the job looks like it failed on whatever the *runtime logs* happen to show — which can be a stale, already-known-fixed error from the **previous** crashed container (e.g. an old `NULLS NOT DISTINCT` migration failure that was already fixed on that same commit), not evidence that the current deploy is broken. `printRuntimeLogs()` passes `data` explicitly and so was unaffected — the giveaway that only the two `data`-less call sites were failing. Fix: `callCaproverApi` must always append `--data <json>`, defaulting to `{}` when the caller passes none.
+- **Don't add redirect shims for moved/removed routes.** When a page or route is relocated or removed, do not leave a redirect stub at the old path — redirecting old routes is never necessary here. Just delete the old route and update any in-app links (`ConfigCard` hrefs, `router.push`/`<Link>` targets) to point at the new location. Deep-linking to sub-views is done with query params on the surviving route (e.g. the consolidated Suggestions admin uses `?tab=categories`/`?tab=import-export`), not separate routes.
 - **Record expensive learnings here.** When a task surfaces an important or expensive-to-rediscover learning — a non-obvious gotcha, a tool that's unavailable in an environment, an automated workflow that handles something for you, or a subtle ordering/dependency constraint — add it to this CLAUDE.md (or the relevant directory `agent.md`) so future sessions don't pay the cost of rediscovering it.
 
 ## E2E Testing Rules (CI vs Deploy)
