+++ anything-frontend/src/hooks/useSuggestionCategories.ts
@@ -2,13 +2,7 @@
 
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { apiClient } from "@/lib/apiClient";
-import type {
-  SuggestionCategory,
-  ExportSuggestionCategoriesResponse,
-  ImportSuggestionCategoriesRequest,
-} from "@/lib/api-client/models/index";
-
-export type { ExportSuggestionCategoriesResponse, ImportSuggestionCategoriesRequest };
+import type { SuggestionCategory } from "@/lib/api-client/models/index";
 
 export function useSuggestionCategories() {
   return useQuery({
@@ -66,31 +60,3 @@ export function useReorderSuggestionCategories() {
   });
 }
 
-export function useExportSuggestionCategories() {
-  return useMutation({
-    mutationFn: async () => {
-      const data = await apiClient.api.suggestionCategories.exportEscaped.get();
-      if (!data) throw new Error("Export failed");
-      const blob = new Blob([JSON.stringify(data, null, 2)], { type: "application/json" });
-      const url = URL.createObjectURL(blob);
-      const anchor = document.createElement("a");
-      anchor.href = url;
-      anchor.download = "suggestion-categories.json";
-      anchor.click();
-      URL.revokeObjectURL(url);
-    },
-  });
-}
-
-export function useImportSuggestionCategories() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async (data: ImportSuggestionCategoriesRequest) => {
-      await apiClient.api.suggestionCategories.importEscaped.post(data);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["suggestionCategories"] });
-    },
-  });
-}
